# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"DetailedError": {
				"properties": {
					"message": {
						"type": "string"
					},
					"standardizedErrorCode": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"standardizedErrorCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaymentInitializeResponse": {
				"properties": {
					"paymentProviderRedirectUrl": {
						"type": "string"
					},
					"paymentReference": {
						"type": "string"
					},
					"accessCode": {
						"type": "string"
					}
				},
				"required": [
					"paymentProviderRedirectUrl",
					"paymentReference",
					"accessCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IServerResponse_PaymentInitializeResponse_": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/PaymentInitializeResponse"
					},
					"error": {
						"type": "string"
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/DetailedError"
						},
						"type": "array"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentInitializeVariant": {
				"enum": [
					"FUND_MAIN_WALLET"
				],
				"type": "string"
			},
			"PaymentInitialize": {
				"properties": {
					"paymentVariant": {
						"$ref": "#/components/schemas/PaymentInitializeVariant"
					},
					"amountMajor": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"paymentVariant"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Pick_PaystackDedicatedNuban.Exclude_keyofPaystackDedicatedNuban.id__": {
				"properties": {
					"uuid": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"dedicatedNubanPayload": {},
					"bankId": {
						"type": "string"
					},
					"bankName": {
						"type": "string"
					},
					"bankAccountNumber": {
						"type": "string"
					},
					"bankAccountName": {
						"type": "string"
					},
					"paystackCustomerId": {
						"type": "string"
					},
					"paystackIntegration": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"uuid",
					"userId",
					"bankId",
					"bankName",
					"bankAccountNumber",
					"bankAccountName",
					"paystackCustomerId",
					"paystackIntegration",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_PaystackDedicatedNuban.id_": {
				"$ref": "#/components/schemas/Pick_PaystackDedicatedNuban.Exclude_keyofPaystackDedicatedNuban.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IServerResponse_Omit_PaystackDedicatedNuban.id__": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Omit_PaystackDedicatedNuban.id_"
					},
					"error": {
						"type": "string"
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/DetailedError"
						},
						"type": "array"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "x-access-token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Wallet",
		"version": "1.0.0",
		"description": "Wallet Service Backend",
		"contact": {
			"name": "Mubarak Aminu"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/payments/paystack/initialize": {
			"post": {
				"operationId": "InitializePaystackPayment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IServerResponse_PaymentInitializeResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PaymentInitialize"
							}
						}
					}
				}
			}
		},
...
		"/api/payments/paystack/dedicated-account": {
			"get": {
				"operationId": "DedicatedAccount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IServerResponse_Omit_PaystackDedicatedNuban.id__"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}